refactor status to handlers
