+++ node/full_test.go
@@ -2,7 +2,6 @@ package node
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"net/http"
@@ -14,8 +13,6 @@ import (
 	"github.com/libp2p/go-libp2p-core/host"
 	"github.com/libp2p/go-libp2p-core/network"
 	"github.com/libp2p/go-libp2p-core/protocol"
-
-	"github.com/celestiaorg/celestia-node/node/rpc/handlers"
 )
 
 func TestNewFull(t *testing.T) {
@@ -143,45 +140,4 @@ type ping struct{}
 func (p ping) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	//nolint:errcheck
 	w.Write([]byte("pong"))
-}
-
-// Should be moved to a specific handlers tests file
-func TestFull_RPCStatus(t *testing.T) {
-	repo := MockRepository(t, DefaultConfig(Full))
-	node, err := New(Full, repo)
-	require.NoError(t, err)
-	require.NotNil(t, node)
-	require.NotNil(t, node.Host)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	t.Cleanup(cancel)
-
-	err = node.Start(ctx)
-	require.NoError(t, err)
-
-	// expected result
-	expectedStatus, err := json.Marshal(
-		handlers.NewStatusMessage(
-			node.Config.P2P.ListenAddresses,
-			node.Config.P2P.Network,
-		),
-	)
-	require.NoError(t, err)
-
-	// get status
-	resp, err := http.Get(fmt.Sprintf("http://%s/status", node.Config.RPC.ListenAddr))
-	require.NoError(t, err)
-
-	// read status
-	actualStatus, err := ioutil.ReadAll(resp.Body)
-	require.NoError(t, err)
-	t.Cleanup(func() {
-		resp.Body.Close()
-	})
-	assert.Equal(t, expectedStatus, actualStatus)
-
-	ctx, cancel = context.WithCancel(context.Background())
-	t.Cleanup(cancel)
-
-	require.NoError(t, node.Stop(ctx))
 }
