+++ node/handlers.go
@@ -1,13 +1,9 @@
-package handlers
+package node
 
 import (
 	"encoding/json"
 	"net/http"
-
-	logging "github.com/ipfs/go-log/v2"
 )
-
-var log = logging.Logger("RPC")
 
 type StatusMessage struct {
 	ListenAddresses []string
